+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.15.9] - in progress
+
+
+- Bug fix in `IterableQueue` when using a multiprocessing SimpleQueue.
+- `mpservice.mpserver.Server.stream` gets a new parameter `preprocess`.
+
+
 ## [0.15.8] - 2024-0921
 
 - Added new class `mpservice.streamer.IterableQueue`.